API Add an AttributesHTML trait
